// src/pages/CreateOperations/components/forms/FormPlannedOperation/constants/enums.ts
export enum FormFieldKey {
  IS_WRITE_OFF = 'isWriteOff',
  IS_SURVICE = 'isService',
  CONTRACTOR = 'contractor',
  PAYMENT_DATE = 'paymentDate',
  ACCRUAL_DATE = 'receiveDate',
  BUDGET_ITEM = 'budgetItem',
  AMOUNT = 'amount',
  NAME = 'name',
  PAYMENT_TYPE = 'paymentType',
}

export enum PaymentType {
  Cashless = 'cashless',
  Cash = 'cash',
}